// src/app/models/attachment.ts
export class Attachment {
    _id?: string;
    name: string;
    description: string;
    modified?: string;
    size?: string;
    mimeType?: string;
    revision?: string;
    path?: string;
    icon?: string;
    created?: string;
    updated?: string;
    createdBy?: string;
    updatedBy?: string;
}